requirements.txt
python-telegram-bot[job-queue]>=21.0
python-dotenv>=1.0.0
pytz>=2024.1
aiofiles>=23.2.1
anthropic>=0.40.0
pyautogui>=0.9.54
pillow>=10.0.0
psutil>=5.9.0
google-generativeai>=0.8.0
